chore(comments): move shared comment components out of scenes
The directory never held a scene. It has been the Discussions side panel's
component tree since it was created, and is now shared by the side panel,
notebooks, session replay, AI observability, logs, and conversations.

Move it to the app-shared layer and repoint the ten importers. Also fix the
two kea paths, which pointed at scenes/comments and scenes/notebooks.
