+++ frontend/src/layout/navigation-3000/sidepanel/panels/discussion/SidePanelDiscussion.tsx
@@ -4,11 +4,11 @@ import { useEffect } from 'react'
 import { IconChat } from '@posthog/icons'
 
 import { humanizeScope } from 'lib/components/ActivityLog/humanizeActivity'
+import { CommentComposer } from 'lib/components/Comments/CommentComposer'
+import { CommentsList } from 'lib/components/Comments/CommentsList'
+import { CommentsLogicProps, commentsLogic } from 'lib/components/Comments/commentsLogic'
 import { WarningHog } from 'lib/components/hedgehogs'
 import { IconWithCount } from 'lib/lemon-ui/icons'
-import { CommentComposer } from 'scenes/comments/CommentComposer'
-import { CommentsList } from 'scenes/comments/CommentsList'
-import { CommentsLogicProps, commentsLogic } from 'scenes/comments/commentsLogic'
 
 import { SidePanelPaneHeader } from '../../components/SidePanelPaneHeader'
 import { SidePanelContentContainer } from '../../SidePanelContentContainer'
