+++ products/conversations/frontend/scenes/ticket/ThreadDiscussions.tsx
@@ -4,12 +4,12 @@ import { useEffect } from 'react'
 import { IconChat } from '@posthog/icons'
 import { LemonTag } from '@posthog/lemon-ui'
 
+import { getCommentAuthorName, getText } from 'lib/components/Comments/Comment'
+import { CommentWithRepliesType, commentsLogic } from 'lib/components/Comments/commentsLogic'
 import { TZLabel } from 'lib/components/TZLabel'
 import { IconSlack } from 'lib/lemon-ui/icons'
 import { ProfileBubbles } from 'lib/lemon-ui/ProfilePicture/ProfileBubbles'
 import { pluralize } from 'lib/utils/strings'
-import { getCommentAuthorName, getText } from 'scenes/comments/Comment'
-import { CommentWithRepliesType, commentsLogic } from 'scenes/comments/commentsLogic'
 
 import { sidePanelStateLogic } from '~/layout/navigation-3000/sidepanel/sidePanelStateLogic'
 import { ActivityScope, CommentType, SidePanelTab } from '~/types'
